# generated/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/register_buckets":{"post":{"summary":"Register Buckets","operationId":"register_buckets","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterBucketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/start_create_bucket":{"post":{"summary":"Start Create Bucket","operationId":"start_create_bucket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBucketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBucketResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/complete_create_bucket":{"patch":{"summary":"Complete Create Bucket","operationId":"complete_create_bucket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBucketIsCompleted"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/start_delete_bucket":{"post":{"summary":"Start Delete Bucket","operationId":"start_delete_bucket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteBucketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteBucketResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/complete_delete_bucket":{"patch":{"summary":"Complete Delete Bucket","operationId":"complete_delete_bucket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteBucketIsCompleted"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/locate_bucket":{"post":{"summary":"Locate Bucket","description":"Given the bucket name, return one or zero physical bucket locators.","operationId":"locate_bucket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocateBucketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocateBucketResponse"}}}},"404":{"description":"Bucket not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/list_buckets":{"post":{"summary":"List Buckets","operationId":"list_buckets","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List Buckets List Buckets Post","type":"array","items":{"$ref":"#/components/schemas/BucketResponse"}}}}}}}},"/head_bucket":{"post":{"summary":"Head Bucket","operationId":"head_bucket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HeadBucketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/put_bucket_versioning":{"post":{"summary":"Put Bucket Versioning","operationId":"put_bucket_versioning","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutBucketVersioningRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Put Bucket Versioning Put Bucket Versioning Post","type":"array","items":{"$ref":"#/components/schemas/LocateBucketResponse"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/check_version_setting":{"post":{"summary":"Check Version Setting","operationId":"check_version_setting","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HeadBucketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Check Version Setting Check Version Setting Post","type":"boolean"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/locate_bucket_status":{"post":{"summary":"Locate Bucket Status","description":"Given the bucket name, return physical bucket status. Currently only used for testing metadata cleanup","operationId":"locate_bucket_status","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocateBucketRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BucketStatus"}}}},"404":{"description":"Bucket not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/update_policy":{"post":{"summary":"Update Policy","operationId":"update_policy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetPolicyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/start_delete_objects":{"post":{"summary":"Start Delete Objects","operationId":"start_delete_objects","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteObjectsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteObjectsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/complete_delete_objects":{"patch":{"summary":"Complete Delete Objects","operationId":"complete_delete_objects","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteObjectsIsCompleted"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/start_warmup":{"post":{"summary":"Start Warmup","description":"Given the logical object information and warmup regions, return one or zero physical object locators.","operationId":"start_warmup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartWarmupRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartWarmupResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/list_objects":{"post":{"summary":"List Objects","operationId":"list_objects","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListObjectRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List Objects List Objects Post","type":"array","items":{"$ref":"#/components/schemas/ObjectResponse"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/list_objects_versioning":{"post":{"summary":"List Objects Versioning","operationId":"list_objects_versioning","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListObjectRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List Objects Versioning List Objects Versioning Post","type":"array","items":{"$ref":"#/components/schemas/ObjectResponse"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/head_object":{"post":{"summary":"Head Object","operationId":"head_object","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HeadObjectRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HeadObjectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/list_multipart_uploads":{"post":{"summary":"List Multipart Uploads","operationId":"list_multipart_uploads","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListObjectRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List Multipart Uploads List Multipart Uploads Post","type":"array","items":{"$ref":"#/components/schemas/MultipartResponse"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/list_parts":{"post":{"summary":"List Parts","operationId":"list_parts","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListPartsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response List Parts List Parts Post","type":"array","items":{"$ref":"#/components/schemas/LogicalPartResponse"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/locate_object_status":{"post":{"summary":"Locate Object Status","description":"Given the logical object information, return the status of the object.\nCurrently only used for testing metadata cleanup.","operationId":"locate_object_status","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocateObjectRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ObjectStatus","title":"Response Locate Object Status Locate Object Status Post","type":"array","items":{"$ref":"#/components/schemas/ObjectStatus"}}}}},"404":{"description":"Object not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/update_metrics":{"post":{"summary":"Update Metrics","operationId":"update_metrics","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Metrics"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/set_multipart_id":{"patch":{"summary":"Set Multipart Id","operationId":"set_multipart_id","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchUploadMultipartUploadId"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/append_part":{"patch":{"summary":"Append Part","operationId":"append_part","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchUploadMultipartUploadPart"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/continue_upload":{"post":{"summary":"Continue Upload","operationId":"continue_upload","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContinueUploadRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Continue Upload Continue Upload Post","type":"array","items":{"$ref":"#/components/schemas/ContinueUploadResponse"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/start_upload":{"post":{"summary":"Start Upload","description":"Reference \n# https://docs.aws.amazon.com/AmazonS3/latest/userguide/AddingObjectstoVersionSuspendedBuckets.html\n# https://docs.aws.amazon.com/AmazonS3/latest/userguide/AddingObjectstoVersioningEnabledBuckets.html\n\nTwo possible operations: \n1. First Write: create new logical objects \n2. Pull on Read: reuse existing logical objects \n3. Copy: create new logical objects in dst locations","operationId":"start_upload","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartUploadRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartUploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/complete_upload":{"patch":{"summary":"Complete Upload","operationId":"complete_upload","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchUploadIsCompleted"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/locate_object":{"post":{"summary":"Locate Object","description":"Given the logical object information, return one or zero physical object locators.","operationId":"locate_object","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocateObjectRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocateObjectResponse"}}}},"404":{"description":"Object not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/healthz":{"get":{"summary":"Healthz","operationId":"healthz","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthcheckResponse"}}}}}}}},"components":{"schemas":{"BucketResponse":{"title":"BucketResponse","required":["bucket","creation_date"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"creation_date":{"title":"Creation Date","type":"string","format":"date-time"}}},"BucketStatus":{"title":"BucketStatus","required":["status"],"type":"object","properties":{"status":{"$ref":"#/components/schemas/Status"}}},"Configuration":{"title":"Configuration","type":"object","properties":{"bucket_name":{"title":"Bucket Name","type":"string","default":"default-skybucket"},"physical_locations":{"title":"Physical Locations","type":"array","items":{"$ref":"#/components/schemas/PhysicalLocation"}}}},"ContinueUploadPhysicalPart":{"title":"ContinueUploadPhysicalPart","required":["part_number","etag"],"type":"object","properties":{"part_number":{"title":"Part Number","type":"integer"},"etag":{"title":"Etag","type":"string"}}},"ContinueUploadRequest":{"title":"ContinueUploadRequest","required":["bucket","key","client_from_region","multipart_upload_id"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"client_from_region":{"title":"Client From Region","type":"string"},"version_id":{"title":"Version Id","type":"integer"},"ttl":{"title":"Ttl","type":"integer"},"multipart_upload_id":{"title":"Multipart Upload Id","type":"string"},"do_list_parts":{"title":"Do List Parts","type":"boolean","default":false},"copy_src_bucket":{"title":"Copy Src Bucket","type":"string"},"copy_src_key":{"title":"Copy Src Key","type":"string"}}},"ContinueUploadResponse":{"title":"ContinueUploadResponse","required":["id","tag","cloud","bucket","region","key","multipart_upload_id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"tag":{"title":"Tag","type":"string"},"cloud":{"title":"Cloud","type":"string"},"bucket":{"title":"Bucket","type":"string"},"region":{"title":"Region","type":"string"},"key":{"title":"Key","type":"string"},"version_id":{"title":"Version Id","type":"string"},"version":{"title":"Version","type":"integer"},"size":{"title":"Size","minimum":0.0,"type":"integer","format":"int64"},"last_modified":{"title":"Last Modified","type":"string","format":"date-time"},"etag":{"title":"Etag","type":"string"},"multipart_upload_id":{"title":"Multipart Upload Id","type":"string"},"ttl":{"title":"Ttl","type":"integer"},"parts":{"title":"Parts","type":"array","items":{"$ref":"#/components/schemas/ContinueUploadPhysicalPart"}},"copy_src_bucket":{"title":"Copy Src Bucket","type":"string"},"copy_src_key":{"title":"Copy Src Key","type":"string"}}},"CreateBucketIsCompleted":{"title":"CreateBucketIsCompleted","required":["id","creation_date"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"creation_date":{"title":"Creation Date","type":"string","format":"date-time"}}},"CreateBucketRequest":{"title":"CreateBucketRequest","required":["bucket","client_from_region"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"client_from_region":{"title":"Client From Region","type":"string"},"warmup_regions":{"title":"Warmup Regions","type":"array","items":{"type":"string"}}}},"CreateBucketResponse":{"title":"CreateBucketResponse","required":["locators"],"type":"object","properties":{"locators":{"title":"Locators","type":"array","items":{"$ref":"#/components/schemas/LocateBucketResponse"}}}},"DeleteBucketIsCompleted":{"title":"DeleteBucketIsCompleted","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"}}},"DeleteBucketRequest":{"title":"DeleteBucketRequest","required":["bucket"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"}}},"DeleteBucketResponse":{"title":"DeleteBucketResponse","required":["locators"],"type":"object","properties":{"locators":{"title":"Locators","type":"array","items":{"$ref":"#/components/schemas/LocateBucketResponse"}}}},"DeleteMarker":{"title":"DeleteMarker","required":["delete_marker"],"type":"object","properties":{"delete_marker":{"title":"Delete Marker","type":"boolean"},"version_id":{"title":"Version Id","type":"string"}}},"DeleteObjectsIsCompleted":{"title":"DeleteObjectsIsCompleted","required":["ids","op_type"],"type":"object","properties":{"ids":{"title":"Ids","type":"array","items":{"type":"integer"}},"multipart_upload_ids":{"title":"Multipart Upload Ids","type":"array","items":{"type":"string"}},"op_type":{"title":"Op Type","type":"array","items":{"type":"string"}}}},"DeleteObjectsRequest":{"title":"DeleteObjectsRequest","required":["bucket","object_identifiers"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"object_identifiers":{"title":"Object Identifiers","type":"object","additionalProperties":{"uniqueItems":true,"type":"array","items":{"type":"integer"}}},"multipart_upload_ids":{"title":"Multipart Upload Ids","type":"array","items":{"type":"string"}}}},"DeleteObjectsResponse":{"title":"DeleteObjectsResponse","required":["locators","delete_markers","op_type"],"type":"object","properties":{"locators":{"title":"Locators","type":"object","additionalProperties":{"type":"array","items":{"$ref":"#/components/schemas/LocateObjectResponse"}}},"delete_markers":{"title":"Delete Markers","type":"object","additionalProperties":{"$ref":"#/components/schemas/DeleteMarker"}},"op_type":{"title":"Op Type","type":"object","additionalProperties":{"type":"string"}}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"HeadBucketRequest":{"title":"HeadBucketRequest","required":["bucket"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"}}},"HeadObjectRequest":{"title":"HeadObjectRequest","required":["bucket","key"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"version_id":{"title":"Version Id","type":"integer"}}},"HeadObjectResponse":{"title":"HeadObjectResponse","required":["bucket","key","size","etag","last_modified"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"size":{"title":"Size","minimum":0.0,"type":"integer","format":"int64"},"etag":{"title":"Etag","type":"string"},"last_modified":{"title":"Last Modified","type":"string","format":"date-time"},"version_id":{"title":"Version Id","type":"integer"}}},"HealthcheckResponse":{"title":"HealthcheckResponse","required":["status"],"type":"object","properties":{"status":{"title":"Status","enum":["OK"],"type":"string"}}},"ListObjectRequest":{"title":"ListObjectRequest","required":["bucket"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"prefix":{"title":"Prefix","type":"string"},"start_after":{"title":"Start After","type":"string"},"max_keys":{"title":"Max Keys","type":"integer"}}},"ListPartsRequest":{"title":"ListPartsRequest","required":["bucket","key","upload_id"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"upload_id":{"title":"Upload Id","type":"string"},"part_number":{"title":"Part Number","type":"integer"}}},"LocateBucketRequest":{"title":"LocateBucketRequest","required":["bucket","client_from_region"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"client_from_region":{"title":"Client From Region","type":"string"}}},"LocateBucketResponse":{"title":"LocateBucketResponse","required":["id","tag","cloud","bucket","region"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"tag":{"title":"Tag","type":"string"},"cloud":{"title":"Cloud","type":"string"},"bucket":{"title":"Bucket","type":"string"},"region":{"title":"Region","type":"string"}}},"LocateObjectRequest":{"title":"LocateObjectRequest","required":["bucket","key","client_from_region"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"client_from_region":{"title":"Client From Region","type":"string"},"version_id":{"title":"Version Id","type":"integer"},"ttl":{"title":"Ttl","type":"integer"}}},"LocateObjectResponse":{"title":"LocateObjectResponse","required":["id","tag","cloud","bucket","region","key"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"tag":{"title":"Tag","type":"string"},"cloud":{"title":"Cloud","type":"string"},"bucket":{"title":"Bucket","type":"string"},"region":{"title":"Region","type":"string"},"key":{"title":"Key","type":"string"},"version_id":{"title":"Version Id","type":"string"},"version":{"title":"Version","type":"integer"},"size":{"title":"Size","minimum":0.0,"type":"integer","format":"int64"},"last_modified":{"title":"Last Modified","type":"string","format":"date-time"},"etag":{"title":"Etag","type":"string"},"multipart_upload_id":{"title":"Multipart Upload Id","type":"string"},"ttl":{"title":"Ttl","type":"integer"}}},"LogicalPartResponse":{"title":"LogicalPartResponse","required":["part_number","etag","size"],"type":"object","properties":{"part_number":{"title":"Part Number","type":"integer"},"etag":{"title":"Etag","type":"string"},"size":{"title":"Size","minimum":0.0,"type":"integer","format":"int64"}}},"Metrics":{"title":"Metrics","required":["timestamp","latency","request_region","destination_region","key","size","op"],"type":"object","properties":{"timestamp":{"title":"Timestamp","type":"string"},"latency":{"title":"Latency","type":"number"},"request_region":{"title":"Request Region","type":"string"},"destination_region":{"title":"Destination Region","type":"string"},"key":{"title":"Key","type":"string"},"size":{"title":"Size","minimum":0.0,"type":"integer","format":"int64"},"op":{"title":"Op","type":"string"}}},"MultipartResponse":{"title":"MultipartResponse","required":["bucket","key","upload_id"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"upload_id":{"title":"Upload Id","type":"string"}}},"ObjectResponse":{"title":"ObjectResponse","required":["bucket","key","size"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"size":{"title":"Size","minimum":0.0,"type":"integer","format":"int64"},"etag":{"title":"Etag","type":"string"},"last_modified":{"title":"Last Modified","type":"string","format":"date-time"},"version_id":{"title":"Version Id","type":"integer"}}},"ObjectStatus":{"title":"ObjectStatus","required":["status"],"type":"object","properties":{"status":{"$ref":"#/components/schemas/Status"}}},"PatchUploadIsCompleted":{"title":"PatchUploadIsCompleted","required":["id","size","etag","last_modified"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"size":{"title":"Size","minimum":0.0,"type":"integer","format":"int64"},"etag":{"title":"Etag","type":"string"},"last_modified":{"title":"Last Modified","type":"string","format":"date-time"},"version_id":{"title":"Version Id","type":"string"}}},"PatchUploadMultipartUploadId":{"title":"PatchUploadMultipartUploadId","required":["id","multipart_upload_id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"multipart_upload_id":{"title":"Multipart Upload Id","type":"string"}}},"PatchUploadMultipartUploadPart":{"title":"PatchUploadMultipartUploadPart","required":["id","part_number","etag","size"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"part_number":{"title":"Part Number","type":"integer"},"etag":{"title":"Etag","type":"string"},"size":{"title":"Size","minimum":0.0,"type":"integer","format":"int64"}}},"PhysicalLocation":{"title":"PhysicalLocation","required":["name","cloud","region","bucket"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"cloud":{"title":"Cloud","type":"string"},"region":{"title":"Region","type":"string"},"bucket":{"title":"Bucket","type":"string"},"prefix":{"title":"Prefix","type":"string","default":""},"is_primary":{"title":"Is Primary","type":"boolean","default":false},"need_warmup":{"title":"Need Warmup","type":"boolean","default":false}}},"PutBucketVersioningRequest":{"title":"PutBucketVersioningRequest","required":["bucket","versioning"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"versioning":{"title":"Versioning","type":"boolean"}}},"RegisterBucketRequest":{"title":"RegisterBucketRequest","required":["bucket","config","versioning"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"config":{"$ref":"#/components/schemas/Configuration"},"versioning":{"title":"Versioning","type":"boolean"}}},"SetPolicyRequest":{"title":"SetPolicyRequest","type":"object","properties":{"get_policy":{"title":"Get Policy","type":"string"},"put_policy":{"title":"Put Policy","type":"string"}}},"StartUploadRequest":{"title":"StartUploadRequest","required":["bucket","key","client_from_region","is_multipart"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"client_from_region":{"title":"Client From Region","type":"string"},"version_id":{"title":"Version Id","type":"integer"},"ttl":{"title":"Ttl","type":"integer"},"is_multipart":{"title":"Is Multipart","type":"boolean"},"copy_src_bucket":{"title":"Copy Src Bucket","type":"string"},"copy_src_key":{"title":"Copy Src Key","type":"string"}}},"StartUploadResponse":{"title":"StartUploadResponse","required":["locators","copy_src_buckets","copy_src_keys"],"type":"object","properties":{"locators":{"title":"Locators","type":"array","items":{"$ref":"#/components/schemas/LocateObjectResponse"}},"multipart_upload_id":{"title":"Multipart Upload Id","type":"string"},"copy_src_buckets":{"title":"Copy Src Buckets","type":"array","items":{"type":"string"}},"copy_src_keys":{"title":"Copy Src Keys","type":"array","items":{"type":"string"}}}},"StartWarmupRequest":{"title":"StartWarmupRequest","required":["bucket","key","client_from_region","warmup_regions"],"type":"object","properties":{"bucket":{"title":"Bucket","type":"string"},"key":{"title":"Key","type":"string"},"client_from_region":{"title":"Client From Region","type":"string"},"version_id":{"title":"Version Id","type":"integer"},"ttl":{"title":"Ttl","type":"integer"},"warmup_regions":{"title":"Warmup Regions","type":"array","items":{"type":"string"}}}},"StartWarmupResponse":{"title":"StartWarmupResponse","required":["src_locator","dst_locators"],"type":"object","properties":{"src_locator":{"$ref":"#/components/schemas/LocateObjectResponse"},"dst_locators":{"title":"Dst Locators","type":"array","items":{"$ref":"#/components/schemas/LocateObjectResponse"}}}},"Status":{"title":"Status","enum":["pending","pending_deletion","ready"],"type":"string","description":"An enumeration."},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}